516/requirements.txt
streamlit
spacy
pandas
pdfminer.six
